Move MultiThreadedEnv hook to RLZoo
