+++ src/ReinforcementLearningZoo/src/ReinforcementLearningZoo.jl
@@ -7,5 +7,6 @@ const RLZoo = ReinforcementLearningZoo
 export RLZoo
 
 include("algorithms/algorithms.jl")
+# include("hooks/hooks.jl") # TotalBatchRewardPerEpisode is broken, need to ensure vector copy works!
 
 end # module
