+++ src/ReinforcementLearningZoo/test/hooks/hooks.jl
@@ -1,0 +1,1 @@
+include("total_batch_reward_per_episode.jl")
