+++ src/ReinforcementLearningZoo/test/runtests.jl
@@ -2,4 +2,5 @@ using Test
 
 @testset "ReinforcementLearningZoo.jl" begin
     # include("cfr/cfr.jl")
+    # include("hooks.jl")
 end
